+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/rng/distribution/Distribution.java
@@ -193,7 +193,7 @@ public interface Distribution {
      * @return an array representing the random sample
      * @throws org.apache.commons.math3.exception.NotStrictlyPositiveException if {@code sampleSize} is not positive
      */
-    double[] sample(int sampleSize);
+    double[] sample(long sampleSize);
 
     /**
      * Sample the given shape
